requirements.txt
boto3==1.35.14
botocore==1.35.14
certifi==2025.10.5
charset-normalizer==3.4.4
cloudpickle==3.1.1
flake8==7.0.0
idna==3.11
jmespath==1.0.1
joblib==1.5.2
llvmlite==0.45.1
matplotlib
numba==0.62.1
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pandas_market_calendars==4.4.0
patsy==1.0.2
pydantic-settings
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
requests==2.32.5
s3transfer==0.10.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
statsmodels==0.14.5
stumpy==1.13.0
sympy
threadpoolctl==3.6.0
tqdm==4.67.1
tsfresh==0.21.1
tzdata==2025.2
urllib3==2.5.0
wheel==0.42.0
XlsxWriter==3.1.9
